refactor: move BaseAnalysisBlock to /analysis
