+++ lightwood/analysis/base.py
